Rename to dgmr package instead
