+++ README.md
@@ -1,5 +1,5 @@
 # Skillful-Nowcasting
-Implementation of DeepMind's Skillful Nowcasting GAN (https://arxiv.org/abs/2104.00954) in PyTorch Lightning.
+Implementation of DeepMind's Skillful Nowcasting GAN Deep Generative Model of Radar (DGMR) (https://arxiv.org/abs/2104.00954) in PyTorch Lightning.
 
 ## Installation
 
@@ -9,4 +9,4 @@ pip install -r requirements.txt
 pip install -e .
 ````
 
-Alternatively, you can also install a usually older version through ```pip install skillful-nowcasting```
+Alternatively, you can also install through ```pip install dgmr```
