+++ dgmr/generators.py
@@ -4,8 +4,8 @@ import torch.nn.functional as F
 from torch.nn.modules.pixelshuffle import PixelShuffle
 from torch.nn.utils.parametrizations import spectral_norm
 from typing import List
-from skillful_nowcasting.common import GBlock, UpsampleGBlock
-from skillful_nowcasting.layers import ConvGRU
+from dgmr.common import GBlock, UpsampleGBlock
+from dgmr.layers import ConvGRU
 from huggingface_hub import PyTorchModelHubMixin
 import logging
 
@@ -20,6 +20,7 @@ class Sampler(torch.nn.Module, PyTorchModelHubMixin):
         latent_channels: int = 768,
         context_channels: int = 384,
         output_channels: int = 1,
+        **kwargs
     ):
         """
         Sampler from the Skillful Nowcasting, see https://arxiv.org/pdf/2104.00954.pdf
@@ -31,7 +32,14 @@ class Sampler(torch.nn.Module, PyTorchModelHubMixin):
             latent_channels: Number of input channels to the lowest ConvGRU layer
         """
         super().__init__()
-        self.forecast_steps = forecast_steps
+        config = locals()
+        config.pop("__class__")
+        config.pop("self")
+        self.config = kwargs.get("config", config)
+        self.forecast_steps = self.config["forecast_steps"]
+        latent_channels = self.config["latent_channels"]
+        context_channels = self.config["context_channels"]
+        output_channels = self.config["output_channels"]
 
         self.convGRU1 = ConvGRU(
             input_channels=latent_channels + context_channels,
